code/data-platform/core/orchestration/dagster/requirements.txt
dagster
dagit
dagster-graphql
dagster-postgres
dagster-duckdb
dagster-docker
dlt[duckdb]
dbt-core
dbt-duckdb
python-dateutil
requests